// src/types.ts
export interface EarthVisualizationProps {
    width?: number | string;
    height?: number | string;
    rotationSpeed?: number;
    numStars?: number;
    earthTexturePath?: string;
    specularMapPath?: string;
    bumpMapPath?: string;
    cloudsTexturePath?: string;
    cloudsAlphaMapPath?: string;
    lightsTexturePath?: string;
}